Moving agnostic methods back into QInterface
