+++ include/qinterface.hpp
@@ -110,6 +110,15 @@ protected:
     virtual real1 Rand() { return rand_distribution(*rand_generator); }
     virtual void SetRandomSeed(uint32_t seed) { rand_generator->seed(seed); }
 
+    virtual void Apply2x2(bitCapInt offset1, bitCapInt offset2, const complex* mtrx, const bitLenInt bitCount,
+        const bitCapInt* qPowersSorted, bool doCalcNorm) { throw "Apply2x2 not implemented in interface"; };
+    virtual void ApplyControlled2x2(bitLenInt control, bitLenInt target, const complex* mtrx, bool doCalcNorm);
+    virtual void ApplyAntiControlled2x2(bitLenInt control, bitLenInt target, const complex* mtrx, bool doCalcNorm);
+    virtual void ApplyDoublyControlled2x2(
+        bitLenInt control1, bitLenInt control2, bitLenInt target, const complex* mtrx, bool doCalcNorm);
+    virtual void ApplyDoublyAntiControlled2x2(
+        bitLenInt control1, bitLenInt control2, bitLenInt target, const complex* mtrx, bool doCalcNorm);
+
 public:
     QInterface(bitLenInt n, std::shared_ptr<std::default_random_engine> rgp = nullptr)
         : rand_distribution(0.0, 1.0)
@@ -270,14 +279,14 @@ public:
      * If float rounding from the application of the matrix might change the state vector norm, "doCalcNorm" should be
      * set to true.
      */
-    virtual void ApplySingleBit(const complex* mtrx, bool doCalcNorm, bitLenInt qubitIndex) = 0;
+    virtual void ApplySingleBit(const complex* mtrx, bool doCalcNorm, bitLenInt qubitIndex);
 
     /**
      * Doubly-controlled NOT gate
      *
      * If both controls are set to 1, the target bit is NOT-ed or X-ed.
      */
-    virtual void CCNOT(bitLenInt control1, bitLenInt control2, bitLenInt target) = 0;
+    virtual void CCNOT(bitLenInt control1, bitLenInt control2, bitLenInt target);
 
     /**
      * Anti doubly-controlled NOT gate
@@ -291,7 +300,7 @@ public:
      *
      * If the control is set to 1, the target bit is NOT-ed or X-ed.
      */
-    virtual void CNOT(bitLenInt control, bitLenInt target) = 0;
+    virtual void CNOT(bitLenInt control, bitLenInt target);
 
     /**
      * Anti controlled NOT gate
@@ -305,7 +314,7 @@ public:
      *
      * Applies a Hadamard gate on qubit at "qubitIndex."
      */
-    virtual void H(bitLenInt qubitIndex) = 0;
+    virtual void H(bitLenInt qubitIndex);
 
     /**
      * Measurement gate
@@ -362,7 +371,7 @@ public:
      * Applies the Pauli "X" operator to the qubit at "qubitIndex." The Pauli
      * "X" operator is equivalent to a logical "NOT."
      */
-    virtual void X(bitLenInt qubitIndex) = 0;
+    virtual void X(bitLenInt qubitIndex);
 
     /**
      * Y gate
@@ -371,7 +380,7 @@ public:
      * "Y" operator is similar to a logical "NOT" with permutation phase
      * effects.
      */
-    virtual void Y(bitLenInt qubitIndex) = 0;
+    virtual void Y(bitLenInt qubitIndex);
 
     /**
      * Z gate
@@ -379,7 +388,7 @@ public:
      * Applies the Pauli "Z" operator to the qubit at "qubitIndex." The Pauli
      * "Z" operator reverses the phase of |1> and leaves |0> unchanged.
      */
-    virtual void Z(bitLenInt qubitIndex) = 0;
+    virtual void Z(bitLenInt qubitIndex);
 
     /**
      * Controlled Y gate
@@ -387,7 +396,7 @@ public:
      * If the "control" bit is set to 1, then the Pauli "Y" operator is applied
      * to "target."
      */
-    virtual void CY(bitLenInt control, bitLenInt target) = 0;
+    virtual void CY(bitLenInt control, bitLenInt target);
 
     /**
      * Controlled Z gate
@@ -395,7 +404,7 @@ public:
      * If the "control" bit is set to 1, then the Pauli "Z" operator is applied
      * to "target."
      */
-    virtual void CZ(bitLenInt control, bitLenInt target) = 0;
+    virtual void CZ(bitLenInt control, bitLenInt target);
 
     /** @} */
 
@@ -466,7 +475,7 @@ public:
      *
      * Rotates as \f$ e^{-i*\theta/2} \f$ around |1> state
      */
-    virtual void RT(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void RT(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction phase shift gate
@@ -481,7 +490,7 @@ public:
      *
      * Rotates as \f$ e^{-i*\theta/2} \f$ around Pauli X axis
      */
-    virtual void RX(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void RX(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction X axis rotation gate
@@ -495,7 +504,7 @@ public:
      *
      * Applies \f$ e^{-i*\theta*I} \f$, exponentiation of the identity operator
      */
-    virtual void Exp(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void Exp(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction (identity) exponentiation gate
@@ -509,7 +518,7 @@ public:
      *
      * Applies \f$ e^{-i*\theta*\sigma_x} \f$, exponentiation of the Pauli X operator
      */
-    virtual void ExpX(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void ExpX(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction Pauli X exponentiation gate
@@ -523,7 +532,7 @@ public:
      *
      * Applies \f$ e^{-i*\theta*\sigma_y} \f$, exponentiation of the Pauli Y operator
      */
-    virtual void ExpY(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void ExpY(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction Pauli Y exponentiation gate
@@ -537,7 +546,7 @@ public:
      *
      * Applies \f$ e^{-i*\theta*\sigma_z} \f$, exponentiation of the Pauli Z operator
      */
-    virtual void ExpZ(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void ExpZ(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction Pauli Z exponentiation gate
@@ -551,7 +560,7 @@ public:
      *
      * If "control" is 1, rotates as \f$ e^{-i*\theta/2} \f$ on Pauli x axis.
      */
-    virtual void CRX(real1 radians, bitLenInt control, bitLenInt target) = 0;
+    virtual void CRX(real1 radians, bitLenInt control, bitLenInt target);
 
     /**
      * Controlled dyadic fraction X axis rotation gate
@@ -566,7 +575,7 @@ public:
      *
      * Rotates as \f$ e^{-i*\theta/2} \f$ around Pauli y axis.
      */
-    virtual void RY(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void RY(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction Y axis rotation gate
@@ -582,7 +591,7 @@ public:
      * If "control" is set to 1, rotates as \f$ e^{-i*\theta/2} \f$ around
      * Pauli Y axis.
      */
-    virtual void CRY(real1 radians, bitLenInt control, bitLenInt target) = 0;
+    virtual void CRY(real1 radians, bitLenInt control, bitLenInt target);
 
     /**
      * Controlled dyadic fraction y axis rotation gate
@@ -597,7 +606,7 @@ public:
      *
      * Rotates as \f$ e^{-i*\theta/2} \f$ around Pauli Z axis.
      */
-    virtual void RZ(real1 radians, bitLenInt qubitIndex) = 0;
+    virtual void RZ(real1 radians, bitLenInt qubitIndex);
 
     /**
      * Dyadic fraction Z axis rotation gate
@@ -613,7 +622,7 @@ public:
      * If "control" is set to 1, rotates as \f$ e^{-i*\theta/2} \f$ around
      * Pauli Zaxis.
      */
-    virtual void CRZ(real1 radians, bitLenInt control, bitLenInt target) = 0;
+    virtual void CRZ(real1 radians, bitLenInt control, bitLenInt target);
 
     /**
      * Controlled dyadic fraction Z axis rotation gate
@@ -630,7 +639,7 @@ public:
      * \f$ around |1> state.
      */
 
-    virtual void CRT(real1 radians, bitLenInt control, bitLenInt target) = 0;
+    virtual void CRT(real1 radians, bitLenInt control, bitLenInt target);
 
     /**
      * Controlled dyadic fraction "phase shift gate"
@@ -1111,7 +1120,7 @@ public:
         bitLenInt valueLength, bitLenInt carryIndex, unsigned char* values) = 0;
 
     /** Swap values of two bits in register */
-    virtual void Swap(bitLenInt qubitIndex1, bitLenInt qubitIndex2) = 0;
+    virtual void Swap(bitLenInt qubitIndex1, bitLenInt qubitIndex2);
 
     /** Bitwise swap */
     virtual void Swap(bitLenInt start1, bitLenInt start2, bitLenInt length);
