--- src/qengine/README.md
@@ -1,5 +1,0 @@
-This directory contains implementations of gates, operators, and rotational
-gates that are largely implemented in terms of either other gates, or
-`QUnit::Apply2x2()`.
-
-Methods that directly modify the state are implemented in `state/`.
