+++ src/qengine/opencl.cpp
