+++ src/qengine/opencl_multi.cpp
