+++ src/qengine/qheader_double.cl
